# Improve pty host startup marks/traces, revive terminal procs in parallel